+++ packages/@expo/cli/e2e/__tests__/install-test.ts
@@ -216,3 +216,40 @@ it(
   // Could take 45s depending on how fast npm installs
   60 * 1000
 );
+
+describe.only('expo-router integration', () => {
+  it(
+    'runs `npx expo install --fix`',
+    async () => {
+      const projectRoot = await setupTestProjectWithOptionsAsync(
+        'install-expo-router-integration',
+        'with-router',
+        {
+          reuseExisting: false,
+          sdkVersion: '52.0.0',
+        }
+      );
+      const pkg = new JsonFile(path.resolve(projectRoot, 'package.json'));
+
+      // Add a package that requires "fixing" when using canary
+      await execa('node', [bin, 'install', '@react-navigation/native@6.1.18'], {
+        cwd: projectRoot,
+      });
+
+      // Ensure `@react-navigation/native` is installed
+      expect(pkg.read().dependencies).toMatchObject({
+        '@react-navigation/native': '6.1.18',
+      });
+
+      // Add `expo@canary` to the project, and `--fix` project dependencies
+      await execa('node', [bin, 'install', '--fix'], { cwd: projectRoot });
+
+      // Ensure `@react-navigation/native` was updated
+      expect(pkg.read().dependencies).toMatchObject({
+        '@react-navigation/native': '^7.0.0',
+      });
+    },
+    // Could take 45s depending on how fast npm installs
+    60 * 1000
+  );
+});